+++ src/climatevision/generator/transport2030/energy_demand/investmentaction.py
@@ -10,7 +10,6 @@ from ...utils import div
 class InvestmentAction:
     """For some transport mechanism additional investments are needed."""
 
-    # Used by road_gds_mhd_action_wire, other_foot_action_infra, other_cycl_action_infra, rail_action_invest_infra, rail_action_invest_station, road_bus_action_infra
     cost_wage: float
     demand_emplo: float
     demand_emplo_new: float
@@ -156,7 +155,6 @@ class InvestmentAction:
 
 @dataclass(kw_only=True)
 class RoadInvestmentAction(InvestmentAction):
-    # Used by road_action_charger
     base_unit: float
 
     @classmethod
