+++ src/climatevision/generator/transport2030/energy_demand/rail.py
@@ -15,7 +15,7 @@ from .investmentaction import InvestmentAction
 class RailPeople:
     LIFT_INTO_RESULT_DICT = ["transport"]
     transport: Transport
-    # Used by rail_ppl_distance and rail_ppl_metro
+
     base_unit: float
     cost_wage: float
     demand_emplo_new: float
@@ -182,7 +182,6 @@ class RailPeople:
 
 @dataclass(kw_only=True)
 class RailPeopleMetroActionInfra:
-    # Used by rail_ppl_metro_action_infra
     base_unit: float
     cost_wage: float
     demand_emplo: float
@@ -238,7 +237,7 @@ class RailPeopleMetroActionInfra:
 class RailPeopleSum:
     LIFT_INTO_RESULT_DICT = ["transport"]
     transport: Transport
-    # Used by rail_ppl
+
     base_unit: float
     cost_wage: float
     demand_emplo_new: float
@@ -317,7 +316,7 @@ class RailPeopleSum:
 class RailGoods:
     LIFT_INTO_RESULT_DICT = ["transport"]
     transport: Transport
-    # Used by rail_gds
+
     base_unit: float
     cost_wage: float
     demand_emplo: float
@@ -400,7 +399,7 @@ class RailGoods:
 class Rail:
     LIFT_INTO_RESULT_DICT = ["transport"]
     transport: Transport
-    # Used by rail
+
     base_unit: float
     cost_wage: float
     demand_emplo_new: float
