+++ src/climatevision/generator/transport2030/energy_demand/road.py
@@ -28,7 +28,7 @@ class TransportInvestments:
 class Road:
     LIFT_INTO_RESULT_DICT = ["transport"]
     transport: Transport
-    # Used by road_gds_ldt_it_ot, road_gds_ldt_ab, road_gds_mhd_it_ot, road_gds_mhd_ab
+
     mileage: float
 
     @classmethod
@@ -412,7 +412,6 @@ class Road:
 
 @dataclass(kw_only=True)
 class RoadCar(Road):
-    # Used by road_car
     LIFT_INTO_RESULT_DICT = ["transport", "fleet_modernisation_cost"]
     fleet_modernisation_cost: TransportInvestments
 
@@ -465,7 +464,6 @@ class BusInvestments(TransportInvestments):
 
 @dataclass(kw_only=True)
 class RoadBus(Road, BusInvestments):
-    # Used by road_bus
     @staticmethod
     def calc_action_infra(
         inputs: Inputs, *, bus_transport_capacity_pkm: float
@@ -592,7 +590,6 @@ class RoadBus(Road, BusInvestments):
 
 @dataclass(kw_only=True)
 class RoadPeople(Road):
-    # Used by road_ppl
     base_unit: float
     cost_wage: float
     demand_emplo_new: float
@@ -648,7 +645,6 @@ class RoadPeople(Road):
 
 @dataclass(kw_only=True)
 class RoadGoodsMediumAndHeavyDuty(Road):
-    # Used by road_gds_mhd
     base_unit: float
     demand_emplo_new: float
     demand_emplo: float
@@ -723,7 +719,6 @@ class RoadGoodsMediumAndHeavyDuty(Road):
 
 @dataclass(kw_only=True)
 class RoadGoodsLightDuty(Road):
-    # Used by road_gds_ldt
     base_unit: float
 
     invest_pa: float
@@ -765,7 +760,6 @@ class RoadGoodsLightDuty(Road):
 
 @dataclass(kw_only=True)
 class RoadGoods(Road):
-    # Used by road_gds
     base_unit: float
     demand_emplo_new: float
     demand_emplo: float
@@ -811,8 +805,6 @@ class RoadGoods(Road):
 
 @dataclass(kw_only=True)
 class RoadSum(Road):
-    # Used by road
-
     demand_emplo_new: float
     demand_emplo: float
     cost_wage: float
